requirements.txt
pandas==2.2.3
folium==0.18.0
opencage==3.0.2
requests==2.32.3
python-dotenv==1.0.1
dash==2.18.2
dash-bootstrap-components==1.6.0